Fix : Repaired ClientOnly Mode
