+++ src/proxy_modes/client_only/client.rs
@@ -1,39 +1,14 @@
-use crate::network::connection::Connection;
-use crate::network::packet::{Packet, PacketCodec};
-use crate::protocol::minecraft::java::handshake::ServerBoundHandshake;
-use crate::protocol::minecraft::java::login::clientbound_loginsuccess::{
-    ClientBoundLoginSuccess, Property,
-};
-use crate::protocol::minecraft::java::login::{
-    ClientBoundEncryptionRequest, ServerBoundEncryptionResponse,
-};
-use crate::protocol::types::{Boolean, Byte, ByteArray, ProtocolString, UnsignedShort, VarInt};
-use crate::server::ServerResponse;
-use crate::version::Version;
-use crate::{EncryptionState, ProtocolRead};
+use std::{error::Error, io};
+
 use async_trait::async_trait;
 use log::{debug, error, info};
 use reqwest::Client;
 use serde::Deserialize;
-use std::io::{self};
 use uuid::Uuid;
 
-use super::{ClientProxyModeHandler, ProxyMessage, ProxyModeMessageType, ServerProxyModeHandler};
-use crate::core::actors::client::MinecraftClient;
-use crate::core::actors::server::{self, MinecraftServer};
-use crate::core::event::MinecraftCommunication;
-use crate::network::connection::PossibleReadValue;
+use crate::{core::{actors::client::MinecraftClient, event::MinecraftCommunication}, network::{connection::PossibleReadValue, packet::{Packet, PacketCodec}}, protocol::{minecraft::java::login::{clientbound_loginsuccess::{ClientBoundLoginSuccess, Property}, ClientBoundEncryptionRequest, ServerBoundEncryptionResponse}, types::{Boolean, ByteArray, ProtocolString, VarInt}}, proxy_modes::{client_only::ClientOnlyMessage, ClientProxyModeHandler}, EncryptionState};
 
-pub struct ClientOnlyMode;
-
-#[derive(Debug)]
-pub enum ClientOnlyMessage {
-    ClientReady(),
-    ServerReady(),
-    ClientLoginAknowledged(Packet),
-
-    ServerThreshold(VarInt),
-}
+use super::ClientOnlyMode;
 
 #[derive(Debug, Deserialize)]
 struct MojangResponse {
@@ -190,7 +165,7 @@ impl ClientProxyModeHandler<MinecraftCommunication<ClientOnlyMessage>> for Clien
         let auth_response = client.get(&url).send().await.map_err(|e| {
             io::Error::new(
                 io::ErrorKind::Other,
-                format!("Mojang API request failed: {}", e),
+                format!("Mojang API request failed: {:?}", e.source()),
             )
         })?;
 
@@ -301,187 +276,4 @@ impl ClientProxyModeHandler<MinecraftCommunication<ClientOnlyMessage>> for Clien
         debug!("Initializing client offline proxy mode");
         Ok(())
     }
-}
-
-#[async_trait]
-impl ServerProxyModeHandler<MinecraftCommunication<ClientOnlyMessage>> for ClientOnlyMode {
-    async fn handle_external_server(
-        &self,
-        data: PossibleReadValue,
-        actor: &mut MinecraftServer<MinecraftCommunication<ClientOnlyMessage>>,
-    ) -> io::Result<()> {
-        match data {
-            PossibleReadValue::Packet(data) => {
-                let _ = actor
-                    .client_sender
-                    .send(MinecraftCommunication::Packet(data))
-                    .await;
-            }
-            _ => {}
-        }
-
-        Ok(())
-    }
-
-    async fn handle_internal_server(
-        &self,
-        message: MinecraftCommunication<ClientOnlyMessage>,
-        actor: &mut MinecraftServer<MinecraftCommunication<ClientOnlyMessage>>,
-    ) -> io::Result<()> {
-        match message {
-            MinecraftCommunication::Packet(data) => {
-                debug!("Received packet from server: {:?}", data);
-                actor
-                    .server_request
-                    .as_mut()
-                    .unwrap()
-                    .server_conn
-                    .as_mut()
-                    .unwrap()
-                    .write_packet(&data)
-                    .await?;
-            }
-            MinecraftCommunication::Shutdown => {
-                debug!("Shutting down server (Received Shutdown message)");
-                actor
-                    .server_request
-                    .as_mut()
-                    .unwrap()
-                    .server_conn
-                    .as_mut()
-                    .unwrap()
-                    .close()
-                    .await?;
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-
-    async fn initialize_server(
-        &self,
-        actor: &mut MinecraftServer<MinecraftCommunication<ClientOnlyMessage>>,
-    ) -> io::Result<()> {
-        debug!("Initializing server offline proxy mode");
-        let mut server_initialised = false;
-
-        if let Some(request) = actor.server_request.as_mut() {
-            if let Some(conn) = request.server_conn.as_mut() {
-                let client_handshake = &request.read_packets[0];
-                let server_addr = conn.peer_addr().await?;
-                let login_start = &request.read_packets[1];
-
-                let server_handshake = prepare_server_handshake(client_handshake, &server_addr)?;
-                conn.write_packet(&server_handshake).await?;
-                conn.write_packet(login_start).await?;
-
-                while !server_initialised {
-                    match conn.read_packet().await? {
-                        packet if packet.id == 0x03 => {
-                            // Set Compression
-                            let threshold = packet.decode::<VarInt>()?;
-                            if threshold.0 >= 0 {
-                                debug!(
-                                    "Received compression config from server, threshold: {}",
-                                    threshold.0
-                                );
-
-                                conn.enable_compression(threshold.0);
-                                match actor
-                                    .client_sender
-                                    .send(MinecraftCommunication::CustomData(
-                                        ClientOnlyMessage::ServerThreshold(threshold),
-                                    ))
-                                    .await
-                                {
-                                    Ok(_) => {}
-                                    Err(e) => {
-                                        error!(
-                                            "Failed to send ServerThreshold message to client: {}",
-                                            e
-                                        );
-                                    }
-                                }
-                            }
-                        }
-                        packet if packet.id == 0x02 => {
-                            // Ignore server's Login Success, we'll use ours
-                            debug!("Received Login Success from server, Server Initialised");
-                            server_initialised = true;
-                        }
-                        packet => {
-                            debug!("Received packet {:?} from server", packet);
-                        }
-                    }
-                }
-
-                match actor
-                    .client_sender
-                    .send(MinecraftCommunication::CustomData(
-                        ClientOnlyMessage::ServerReady(),
-                    ))
-                    .await
-                {
-                    Ok(_) => {}
-                    Err(e) => {
-                        error!("Failed to send ServerReady message to client: {}", e);
-                    }
-                };
-
-                if let Some(msg) = actor.server_receiver.recv().await {
-                    match msg {
-                        MinecraftCommunication::CustomData(
-                            ClientOnlyMessage::ClientLoginAknowledged(packet),
-                        ) => {
-                            conn.write_packet(&packet).await?;
-                        }
-                        _ => {
-                            error!("Unexpected message waited Aknowledge got : {:?}", msg);
-                        }
-                    }
-                }
-            }
-        }
-
-        if let Some(server_request) = &mut actor.server_request {
-            if let Some(server_conn) = &mut server_request.server_conn {
-                for packet in server_request.read_packets.iter() {
-                    server_conn.write_packet(packet).await?;
-                }
-            } else {
-                error!("Server connection is None");
-            }
-        } else {
-            error!("Server request is None");
-        }
-        Ok(())
-    }
-}
-impl ProxyMessage for ClientOnlyMessage {}
-
-impl ProxyModeMessageType for ClientOnlyMode {
-    type Message = ClientOnlyMessage;
-}
-
-fn prepare_server_handshake(
-    client_handshake: &Packet,
-    server_addr: &std::net::SocketAddr,
-) -> io::Result<Packet> {
-    let mut cursor = std::io::Cursor::new(&client_handshake.data);
-    let (protocol_version, _) = VarInt::read_from(&mut cursor)?;
-
-    let server_handshale = ServerBoundHandshake {
-        protocol_version,
-        server_address: ProtocolString(server_addr.ip().to_string()),
-        server_port: UnsignedShort(server_addr.port()),
-        next_state: Byte(2),
-    };
-
-    let handshake = Packet::try_from(&server_handshale).map_err(|e| {
-        io::Error::new(
-            io::ErrorKind::InvalidData,
-            format!("Failed to create server handshake packet: {}", e),
-        )
-    })?;
-    Ok(handshake)
 }
